requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
sqlalchemy==2.0.32
chromadb==0.5.5
sentence-transformers==3.0.1
python-multipart==0.0.9
pydantic==2.8.2
openai==1.106.0
python-dotenv==1.1.1